refactor: more tsx added
