+++ components/Admin/useUsersToSubmit.tsx
@@ -3,15 +3,15 @@ import { useState, useMemo, useCallback } from 'react';
 const useUsersToSubmit = () => {
   const [usersToSubmit, setUsersToSubmit] = useState([]);
 
-  const findInstanceByUserId = (user, userList) => {
+  const findInstanceByUserId = (user: any, userList: Array<any>) => {
     const targetUserInList = userList.find((item) => item.id === user.id);
     return targetUserInList !== undefined;
   };
 
   const addUsersToSubmit = useCallback((updatedUser) => {
-    setUsersToSubmit((prevData) => {
+    setUsersToSubmit((prevData: any) => {
       if (findInstanceByUserId(updatedUser, prevData)) {
-        return prevData.map((row) =>
+        return prevData.map((row: any) =>
           row.id === updatedUser.id ? { ...row, ...updatedUser } : row,
         );
       }
