+++ components/Scrollbar/index.tsx
@@ -1,0 +1,12 @@
+import React, { forwardRef } from 'react';
+import SimpleBar from 'simplebar-react';
+import { styled } from '@mui/material/styles';
+import 'simplebar/dist/simplebar.min.css';
+
+const ScrollbarRoot = styled(SimpleBar)``;
+
+const Scrollbar = forwardRef((props, ref: any) => {
+  return <ScrollbarRoot ref={ref} {...props} />;
+});
+
+export default Scrollbar;
